+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Move tests to actual tests
+- Use `runes.Remove` instead of deprecated `transform.RemoveFunc`
 
 ## v0.1.0